tests, missing: proxy & contract signer

+++ src/core/AllowanceCalldata.sol
@@ -12,10 +12,11 @@ uint256 constant AND = 5;
 uint256 constant OR = 6;
 
 library AllowanceCalldata {
-    function sliceRLPItems(
-        RLPReader.RLPItem[] memory arguments,
-        uint256 start
-    ) internal pure returns (RLPReader.RLPItem[] memory newArguments) {
+    function sliceRLPItems(RLPReader.RLPItem[] memory arguments, uint256 start)
+        internal
+        pure
+        returns (RLPReader.RLPItem[] memory newArguments)
+    {
         newArguments = new RLPReader.RLPItem[](arguments.length - start);
         uint256 initialStart = start;
         for (; start < arguments.length; start++) {
@@ -30,15 +31,11 @@ library AllowanceCalldata {
     ) internal view returns (bool canPass) {
         if (allowedArguments.length == 0) return true;
         for (uint256 i = 0; i < allowedArguments.length; i++) {
-            RLPReader.RLPItem[] memory prefixAndArg = RLPReader.toList(
-                allowedArguments[i]
-            );
+            RLPReader.RLPItem[] memory prefixAndArg = RLPReader.toList(allowedArguments[i]);
             uint256 prefix = RLPReader.toUint(prefixAndArg[0]);
 
             if (prefix == ANY) {} else if (prefix == EQ) {
-                bytes memory allowedArgument = RLPReader.toBytes(
-                    prefixAndArg[1]
-                );
+                bytes memory allowedArgument = RLPReader.toBytes(prefixAndArg[1]);
                 bytes memory argument = RLPReader.toBytes(arguments[i]);
                 canPass = keccak256(allowedArgument) == keccak256(argument);
             } else if (prefix == LT) {
@@ -50,30 +47,16 @@ library AllowanceCalldata {
                 uint256 argument = RLPReader.toUint(arguments[i]);
                 canPass = argument > allowedArgument;
             } else if (prefix == OR) {
-                RLPReader.RLPItem[] memory subAllowance = RLPReader.toList(
-                    prefixAndArg[1]
-                );
-                canPass = validateArguments(
-                    subAllowance,
-                    sliceRLPItems(arguments, i),
-                    true
-                );
+                RLPReader.RLPItem[] memory subAllowance = RLPReader.toList(prefixAndArg[1]);
+                canPass = validateArguments(subAllowance, sliceRLPItems(arguments, i), true);
                 i++;
             } else if (prefix == NE) {
-                bytes memory allowedArgument = RLPReader.toBytes(
-                    prefixAndArg[1]
-                );
+                bytes memory allowedArgument = RLPReader.toBytes(prefixAndArg[1]);
                 bytes memory argument = RLPReader.toBytes(arguments[i]);
                 canPass = keccak256(allowedArgument) != keccak256(argument);
             } else if (prefix == AND) {
-                RLPReader.RLPItem[] memory subAllowance = RLPReader.toList(
-                    prefixAndArg[1]
-                );
-                canPass = validateArguments(
-                    subAllowance,
-                    sliceRLPItems(arguments, i),
-                    false
-                );
+                RLPReader.RLPItem[] memory subAllowance = RLPReader.toList(prefixAndArg[1]);
+                canPass = validateArguments(subAllowance, sliceRLPItems(arguments, i), false);
                 i++;
             } else {
                 revert("Invalid calldata prefix");
@@ -85,15 +68,13 @@ library AllowanceCalldata {
         return canPass;
     }
 
-    function isAllowedCalldata(
-        bytes memory allowed,
-        bytes memory data,
-        uint value
-    ) internal view returns (bool isOk) {
+    function isAllowedCalldata(bytes memory allowed, bytes memory data, uint256 value)
+        internal
+        view
+        returns (bool isOk)
+    {
         RLPReader.RLPItem memory RLPAllowed = RLPReader.toRlpItem(allowed);
-        RLPReader.RLPItem[] memory allowedArguments = RLPReader.toList(
-            RLPAllowed
-        );
+        RLPReader.RLPItem[] memory allowedArguments = RLPReader.toList(RLPAllowed);
         RLPReader.RLPItem memory RLPData = RLPReader.toRlpItem(data);
         RLPReader.RLPItem[] memory arguments = RLPReader.toList(RLPData);
         if (allowedArguments.length != arguments.length) {
@@ -105,16 +86,11 @@ library AllowanceCalldata {
         isOk = validateArguments(allowedArguments, arguments, false);
     }
 
-    function RLPtoABI(
-        bytes memory data
-    ) internal pure returns (bytes memory abiEncoded) {
+    function RLPtoABI(bytes memory data) internal pure returns (bytes memory abiEncoded) {
         RLPReader.RLPItem memory RLPData = RLPReader.toRlpItem(data);
         RLPReader.RLPItem[] memory arguments = RLPReader.toList(RLPData);
         for (uint256 i = 1; i < arguments.length; i++) {
-            abiEncoded = bytes.concat(
-                abiEncoded,
-                RLPReader.toBytes(arguments[i])
-            );
+            abiEncoded = bytes.concat(abiEncoded, RLPReader.toBytes(arguments[i]));
         }
     }
 }
